This is an Arduino camera slider with a panning camera mount.
First attempt at GitHub repository.  The project will be the source code
that drive a camera slider with an optional pan motor drive as well.
Using one stepper motor to the slider drive and a worm drive for the
pan.  I'd like the pan to be able to track with the stars for doing long
exposure night shots.

At this point in time there's still lots to do and figure out.